\name{checkargssim}
\alias{checkargssim}

\title{ For internal use. }
\description{
 For internal use. Check the appropriateness of input arguments of function simplesimint.
}
\usage{
checkargssim(coef, vcov, cmat)
}

\arguments{
  \item{coef}{ a vector of parameters }
  \item{vcov}{ the corresponding variance-covariance matrix }
  \item{cmat}{ a contrast matrix }
}


\keyword{ misc }
