\name{minus2slash}
\alias{minus2slash}

\title{ Replace - by / in character strings}
\description{
  Replace - by / in character strings
}
\usage{
minus2slash(x)
}

\arguments{
  \item{x}{ a vector of character strings }
}

\value{
a vector of character strings
}

\keyword{ misc }