\name{print.CInp}
\alias{print.CInp}

\title{ Print function for CInp }
\description{
 Print fucntion for objects of class CInp.
}
\usage{
\method{print}{CInp}(x, ...)
}

\arguments{
  \item{x}{ an object of class \code{CInp} as can be obtained by calling \code{\link{CInp}} }
  \item{\dots}{ further arguments to be passed to \code{print}}
}

\keyword{print }
