\name{print.SCSnp}
\alias{print.SCSnp}

\title{ Print function for SCSnp }
\description{
 A print function for objects of class "SCSnp"
}
\usage{
\method{print}{SCSnp}(x, ...)
}

\arguments{
  \item{x}{ an object of class \code{SCSnp} as can be obtained by calling \code{\link{SCSnp}} }
  \item{\dots}{ arguments to be passed to \code{print} }
}


\keyword{ print }
