\name{print.simplesimint}
\alias{print.simplesimint}

\title{ Print objects of class "simplesimint" }
\description{
 Print function for objects of class "simplesimint"
}
\usage{
\method{print}{simplesimint}(x, ...)
}

\arguments{
  \item{x}{ an object of class \code{"simplesimint"} as can be obtained by calling \code{\link{simplesimint}} }
  \item{\dots}{ further arguments to be passed to \code{print}}
}

\value{
Prints to console.
}

\seealso{ \code{\link{summary.simplesimint}} }


\keyword{ print }

