\name{print.UnlogCI}
\alias{print.UnlogCI}

\title{ Print function for UnlogCI }
\description{
Print out the results of UnlogCI.
}
\usage{
\method{print}{UnlogCI}(x, ...)
}

\arguments{
  \item{x}{An object of class \code{"UnlogCI"} }
  \item{\dots}{ further arguments to be passed to \code{print} }
}

\value{
A print out.
}

\keyword{ print }
