\name{print.np}
\alias{print.SCSnp}
\alias{print.CInp}
\alias{print.simplesimint}
\alias{print.UnlogCI}

\title{ Print function for SCSnp }
\description{
 A print functions.
}
\usage{
\method{print}{SCSnp}(x, ...)
\method{print}{CInp}(x, ...)
\method{print}{simplesimint}(x, ...)
\method{print}{UnlogCI}(x, ...)
}

\arguments{
  \item{x}{ an object of class \code{SCSnp} as can be obtained by calling \code{\link{SCSnp}} }
  \item{\dots}{ arguments to be passed to \code{print} }
}


\keyword{ internal }
