\name{summary.simplesimint}
\alias{summary.simplesimint}

\title{ Detailed print out for simplesimint objects }
\description{
 Produce a detailed print out for objects of class "simplesimint".
}
\usage{
\method{summary}{simplesimint}(object, ...)
}

\arguments{
  \item{object}{an object of class \code{"simplesimint"} as can be obtained by calling \code{\link{simplesimint}} }
  \item{\dots}{ further arguments to be passed to \code{print} }
}

\value{
Print
}


\keyword{ print }

